feat: 新增雲端同步控制器的單元測試
- 為 DriveCloudSyncController 新增多個單元測試，涵蓋了不同的功能和邊界情況。
- 測試了雲端同步卡片的刷新邏輯，包括從遠端獲取連接狀態和元數據的情況。
- 增加了對於用戶身份初始化失敗、遠端狀態斷開連接等情況的錯誤處理測試。
- 測試了上傳前的確認對話框邏輯，確保在跨安裝情況下用戶能夠選擇是否繼續上傳。
- 確保在不同的狀態下（如已連接、未連接、衝突等）正確渲染雲端同步卡片的顯示。

這些測試的新增旨在提高代碼的穩定性和可維護性，並確保在未來的開發中能夠及早發現潛在的問題。
