+++ tests/unit/options/DriveCloudSyncController.init.test.js
@@ -2,11 +2,13 @@
  * @jest-environment jsdom
  */
 
+jest.mock('../../../pages/options/confirmDialog.js', () => ({
+  confirmDialog: jest.fn().mockResolvedValue(true),
+}));
+
 import {
   initCloudSyncController,
-  setCloudSyncCardVisibility,
   renderCloudSyncCard,
-  refreshCloudSyncCard,
 } from '../../../pages/options/DriveCloudSyncController.js';
 import * as driveClient from '../../../scripts/auth/driveClient.js';
 import { RUNTIME_ACTIONS } from '../../../scripts/config/shared/runtimeActions.js';
@@ -16,166 +18,18 @@ import { ErrorHandler } from '../../../scripts/utils/ErrorHandler.js';
 import { sanitizeApiError } from '../../../scripts/utils/ApiErrorSanitizer.js';
 import { DRIVE_SYNC_ERROR_CODES } from '../../../scripts/config/extension/driveSyncErrorCodes.js';
 
-jest.mock('../../../pages/options/confirmDialog.js', () => ({
-  confirmDialog: jest.fn().mockResolvedValue(true),
-}));
-
-const getConfirmDialogMock = () => require('../../../pages/options/confirmDialog.js').confirmDialog;
-
-async function flushAsyncWork() {
-  await Promise.resolve();
-  await Promise.resolve();
-  await Promise.resolve();
-  await Promise.resolve();
-  await Promise.resolve();
-  await Promise.resolve();
-  await Promise.resolve();
-  await Promise.resolve();
-}
-
-const getStorageString = (key, state) => (Object.hasOwn(state, key) ? { [key]: state[key] } : {});
-
-const getStorageArray = (keys, state) => {
-  const result = {};
-  for (const key of keys) {
-    if (Object.hasOwn(state, key)) {
-      result[key] = state[key];
-    }
-  }
-  return result;
-};
-
-const getStorageObject = (keysObj, state) => {
-  const result = {};
-  for (const [key, defaultVal] of Object.entries(keysObj)) {
-    result[key] = Object.hasOwn(state, key) ? state[key] : defaultVal;
-  }
-  return result;
-};
-
-const getFromStorage = (keys, state) => {
-  if (typeof keys === 'string') {
-    return getStorageString(keys, state);
-  }
-  if (Array.isArray(keys)) {
-    return getStorageArray(keys, state);
-  }
-  if (keys && typeof keys === 'object') {
-    return getStorageObject(keys, state);
-  }
-  return { ...state };
-};
-
-const DEFAULT_DISCONNECTED_CONNECTION = {
-  connected: false,
-  email: null,
-  connectedAt: null,
-};
-
-const DEFAULT_EMPTY_SNAPSHOT = {
-  exists: false,
-  updatedAt: null,
-  size: null,
-  sourceInstallationId: null,
-  sourceProfileId: null,
-};
-
-const REMOTE_CONNECTED_AT = '2026-04-20T00:00:00.000Z';
-const REMOTE_SNAPSHOT_UPDATED_AT = '2026-04-20T09:30:00Z';
-const LOCAL_INSTALLATION_ID = 'local-id-001';
-const REMOTE_INSTALLATION_ID = 'remote-id-999';
-
-const CROSS_INSTALL_SNAPSHOT = {
-  exists: true,
-  updatedAt: 'time',
-  size: null,
-  sourceInstallationId: REMOTE_INSTALLATION_ID,
-  sourceProfileId: null,
-};
-
-function setupCloudSyncDom() {
-  document.body.innerHTML = `
-    <div id="cloud-sync-card">
-      <div id="drive-state-logged-out"></div>
-      <div id="drive-state-disconnected"></div>
-      <div id="drive-state-connected"></div>
-      <div id="drive-state-conflict">
-        <p id="drive-conflict-remote-time"></p>
-      </div>
-      <div id="drive-error-banner">
-        <div id="drive-error-code"></div>
-        <div id="drive-error-time"></div>
-      </div>
-      <div id="drive-loading-overlay">
-        <div id="drive-loading-text"></div>
-      </div>
-      <div id="drive-sync-status"></div>
-      <div id="drive-connected-email"></div>
-      <div id="drive-last-upload-text"></div>
-      <p id="drive-source-warning" hidden></p>
-
-      <select id="drive-frequency-select">
-        <option value="off">Off</option>
-        <option value="daily">Daily</option>
-        <option value="weekly">Weekly</option>
-        <option value="monthly">Every 30 days</option>
-      </select>
-      <output id="drive-auto-sync-status" aria-live="polite">
-        <span id="drive-auto-sync-status-text"></span>
-      </output>
-
-      <button id="drive-connect-button"></button>
-      <button id="drive-login-prompt-button"></button>
-      <button id="drive-upload-button"></button>
-      <button id="drive-download-button"></button>
-      <button id="drive-disconnect-button"></button>
-      <button id="drive-conflict-download-button"></button>
-      <button id="drive-conflict-force-upload-button"></button>
-    </div>
-  `;
-}
-
-function installChromeMock(sendMessage, storageState) {
-  globalThis.chrome = {
-    runtime: {
-      sendMessage,
-    },
-    storage: {
-      local: {
-        get: jest.fn().mockImplementation(async keys => getFromStorage(keys, storageState)),
-        set: jest.fn().mockImplementation(async patch => {
-          Object.assign(storageState, patch);
-        }),
-        remove: jest.fn().mockImplementation(async keys => {
-          const keyList = Array.isArray(keys) ? keys : [keys];
-          for (const key of keyList) {
-            delete storageState[key];
-          }
-        }),
-      },
-    },
-  };
-}
-
-function setupConfirmDialogMock() {
-  getConfirmDialogMock().mockReset();
-  getConfirmDialogMock().mockResolvedValue(true);
-}
-
-function spyOnDriveClientDefaults() {
-  jest.spyOn(driveClient, 'getDriveSyncMetadata');
-  jest.spyOn(driveClient, 'ensureDriveSyncIdentity').mockResolvedValue('local-install');
-  jest.spyOn(driveClient, 'startDriveOAuthFlow').mockResolvedValue();
-  jest.spyOn(driveClient, 'disconnectDrive').mockResolvedValue();
-  jest.spyOn(driveClient, 'clearDriveSyncMetadata').mockResolvedValue();
-  jest.spyOn(driveClient, 'fetchDriveConnectionStatus').mockResolvedValue({
-    ...DEFAULT_DISCONNECTED_CONNECTION,
-  });
-  jest.spyOn(driveClient, 'fetchDriveSnapshotStatus').mockResolvedValue({
-    ...DEFAULT_EMPTY_SNAPSHOT,
-  });
-  jest.spyOn(driveClient, 'setDriveConnection');
-}
+import {
+  flushAsyncWork,
+  DEFAULT_DISCONNECTED_CONNECTION,
+  DEFAULT_EMPTY_SNAPSHOT,
+  REMOTE_CONNECTED_AT,
+  REMOTE_SNAPSHOT_UPDATED_AT,
+  setupCloudSyncDom,
+  installChromeMock,
+  setupConfirmDialogMock,
+  spyOnDriveClientDefaults,
+  getConfirmDialogMock,
+} from './DriveCloudSyncController.shared.js';
 
 describe('DriveCloudSyncController', () => {
   let mockSendMessage;
@@ -198,314 +52,6 @@ describe('DriveCloudSyncController', () => {
     jest.useRealTimers();
     jest.restoreAllMocks();
     delete globalThis.chrome;
-  });
-
-  describe('setCloudSyncCardVisibility', () => {
-    it('shows card if logged in', () => {
-      setCloudSyncCardVisibility(true);
-      expect(document.querySelector('#cloud-sync-card').classList.contains('hidden')).toBe(false);
-    });
-
-    it('hides card if logged out', () => {
-      setCloudSyncCardVisibility(false);
-      expect(document.querySelector('#cloud-sync-card').classList.contains('hidden')).toBe(true);
-    });
-  });
-
-  describe('logged-out state', () => {
-    it('未登入時應顯示提示狀態而不是隱藏整張卡片', async () => {
-      await initCloudSyncController(false);
-
-      expect(document.querySelector('#cloud-sync-card').classList.contains('hidden')).toBe(false);
-      expect(document.querySelector('#drive-state-logged-out').classList.contains('hidden')).toBe(
-        false
-      );
-      expect(document.querySelector('#drive-state-disconnected').classList.contains('hidden')).toBe(
-        true
-      );
-      expect(document.querySelector('#drive-state-connected').classList.contains('hidden')).toBe(
-        true
-      );
-      expect(document.querySelector('#drive-state-conflict').classList.contains('hidden')).toBe(
-        true
-      );
-      expect(document.querySelector('#drive-loading-overlay').classList.contains('hidden')).toBe(
-        true
-      );
-      expect(document.querySelector('#drive-login-prompt-button').disabled).toBe(true);
-    });
-  });
-
-  describe('renderCloudSyncCard', () => {
-    it('renders disconnected state correctly', () => {
-      renderCloudSyncCard({ connectionEmail: null });
-      expect(document.querySelector('#drive-state-disconnected').classList.contains('hidden')).toBe(
-        false
-      );
-      expect(document.querySelector('#drive-state-connected').classList.contains('hidden')).toBe(
-        true
-      );
-      expect(document.querySelector('#drive-state-conflict').classList.contains('hidden')).toBe(
-        true
-      );
-      expect(document.querySelector('#drive-error-banner').classList.contains('hidden')).toBe(true);
-    });
-
-    it('renders connected state correctly without conflict', () => {
-      renderCloudSyncCard({
-        connectionEmail: 'test@notion.so',
-        lastSuccessfulUploadAt: '2023-01-01T00:00:00Z',
-        needsManualReview: false,
-      });
-      expect(document.querySelector('#drive-state-disconnected').classList.contains('hidden')).toBe(
-        true
-      );
-      expect(document.querySelector('#drive-state-connected').classList.contains('hidden')).toBe(
-        false
-      );
-      expect(document.querySelector('#drive-state-conflict').classList.contains('hidden')).toBe(
-        true
-      );
-
-      expect(document.querySelector('#drive-connected-email').textContent).toBe('test@notion.so');
-      expect(document.querySelector('#drive-last-upload-text').textContent).toContain(
-        UI_MESSAGES.CLOUD_SYNC.LAST_UPLOAD_PREFIX
-      );
-      expect(document.querySelector('#drive-last-upload-text').textContent).toContain('（');
-      expect(document.querySelector('#drive-last-upload-text').textContent).toContain('）');
-    });
-
-    it('transientAuthError=true 時應優先顯示暫時登入失效並停用操作按鈕', () => {
-      renderCloudSyncCard(
-        {
-          connectionEmail: 'stale@notion.so',
-          needsManualReview: false,
-        },
-        {
-          transientAuthError: true,
-        }
-      );
-
-      expect(document.querySelector('#drive-state-disconnected').classList.contains('hidden')).toBe(
-        false
-      );
-      expect(document.querySelector('#drive-state-connected').classList.contains('hidden')).toBe(
-        true
-      );
-      expect(document.querySelector('#drive-sync-status').textContent).toContain('臨時登入失效');
-      expect(document.querySelector('#drive-sync-status').textContent).toContain('重新登入');
-      expect(document.querySelector('#drive-sync-status').className).toContain('error');
-      expect(document.querySelector('#drive-upload-button').disabled).toBe(true);
-      expect(document.querySelector('#drive-download-button').disabled).toBe(true);
-      expect(document.querySelector('#drive-disconnect-button').disabled).toBe(true);
-      expect(document.querySelector('#drive-connect-button').disabled).toBe(true);
-    });
-
-    it('renders conflict state correctly', () => {
-      // simulate conflict state
-      renderCloudSyncCard({
-        connectionEmail: 'test@notion.so',
-        needsManualReview: true,
-        lastErrorCode: DRIVE_SYNC_ERROR_CODES.REMOTE_SNAPSHOT_NEWER,
-      });
-      expect(document.querySelector('#drive-state-disconnected').classList.contains('hidden')).toBe(
-        true
-      );
-      expect(document.querySelector('#drive-state-connected').classList.contains('hidden')).toBe(
-        true
-      );
-      expect(document.querySelector('#drive-state-conflict').classList.contains('hidden')).toBe(
-        false
-      );
-
-      // Error banner is hidden because REMOTE_SNAPSHOT_NEWER is considered conflict, not generic error
-      expect(document.querySelector('#drive-error-banner').classList.contains('hidden')).toBe(true);
-    });
-
-    it('conflict state 應在 #drive-conflict-remote-time 顯示帶 prefix 的格式化遠端時間', () => {
-      renderCloudSyncCard({
-        connectionEmail: 'test@notion.so',
-        needsManualReview: true,
-        lastErrorCode: DRIVE_SYNC_ERROR_CODES.REMOTE_SNAPSHOT_NEWER,
-        lastKnownRemoteUpdatedAt: '2026-05-20T08:00:00Z',
-      });
-
-      const remoteTimeText = document.querySelector('#drive-conflict-remote-time').textContent;
-      expect(remoteTimeText).toContain(UI_MESSAGES.CLOUD_SYNC.CONFLICT_REMOTE_PREFIX);
-      expect(remoteTimeText).toContain('（');
-      expect(remoteTimeText).toContain('）');
-    });
-
-    it('conflict state 缺 lastKnownRemoteUpdatedAt 時 #drive-conflict-remote-time 應為空字串', () => {
-      renderCloudSyncCard({
-        connectionEmail: 'test@notion.so',
-        needsManualReview: true,
-        lastErrorCode: DRIVE_SYNC_ERROR_CODES.REMOTE_SNAPSHOT_NEWER,
-      });
-
-      expect(document.querySelector('#drive-conflict-remote-time').textContent).toBe('');
-    });
-
-    it('從 conflict 切回 connected 時 #drive-conflict-remote-time 應被清空避免 stale 文字殘留', () => {
-      renderCloudSyncCard({
-        connectionEmail: 'test@notion.so',
-        needsManualReview: true,
-        lastErrorCode: DRIVE_SYNC_ERROR_CODES.REMOTE_SNAPSHOT_NEWER,
-        lastKnownRemoteUpdatedAt: '2026-05-20T08:00:00Z',
-      });
-      expect(document.querySelector('#drive-conflict-remote-time').textContent).not.toBe('');
-
-      renderCloudSyncCard({
-        connectionEmail: 'test@notion.so',
-        needsManualReview: false,
-        lastSuccessfulUploadAt: '2026-05-21T09:00:00Z',
-      });
-      expect(document.querySelector('#drive-conflict-remote-time').textContent).toBe('');
-    });
-
-    it('renders other generic errors correctly', () => {
-      renderCloudSyncCard({
-        connectionEmail: 'test@notion.so',
-        lastErrorCode: DRIVE_SYNC_ERROR_CODES.UPLOAD_FAILED,
-        lastErrorAt: '2023-01-02T00:00:00Z',
-      });
-      expect(document.querySelector('#drive-error-banner').classList.contains('hidden')).toBe(
-        false
-      );
-      expect(document.querySelector('#drive-error-code').textContent).toContain(
-        DRIVE_SYNC_ERROR_CODES.UPLOAD_FAILED
-      );
-      expect(document.querySelector('#drive-error-time').textContent).toContain(
-        UI_MESSAGES.CLOUD_SYNC.ERROR_TIME_PREFIX
-      );
-    });
-
-    it('falls back to raw timestamp when date formatting throws', () => {
-      const toLocaleStringSpy = jest
-        .spyOn(Date.prototype, 'toLocaleString')
-        .mockImplementation(() => {
-          throw new Error('format failed');
-        });
-
-      renderCloudSyncCard({
-        connectionEmail: 'test@notion.so',
-        lastSuccessfulUploadAt: 'invalid-date-value',
-        lastErrorCode: DRIVE_SYNC_ERROR_CODES.UPLOAD_FAILED,
-        lastErrorAt: 'invalid-error-date',
-      });
-
-      expect(document.querySelector('#drive-last-upload-text').textContent).toContain(
-        'invalid-date-value'
-      );
-      expect(document.querySelector('#drive-error-time').textContent).toContain(
-        'invalid-error-date'
-      );
-
-      toLocaleStringSpy.mockRestore();
-    });
-
-    it('valid timestamp display includes a local timezone label', () => {
-      renderCloudSyncCard({
-        connectionEmail: 'timezone@test.dev',
-        lastKnownRemoteUpdatedAt: '2026-04-19T12:00:00.000Z',
-      });
-
-      const text = document.querySelector('#drive-last-upload-text').textContent;
-      expect(text).toContain('雲端備份：');
-      expect(text).toContain('（');
-      expect(text).toContain('）');
-    });
-
-    it.each([
-      {
-        name: 'lastSuccessfulUploadAt 為 null 但 lastKnownRemoteUpdatedAt 有值 → 顯示「雲端備份」',
-        metadata: {
-          connectionEmail: 'cross-device@test.dev',
-          lastSuccessfulUploadAt: null,
-          lastKnownRemoteUpdatedAt: '2026-04-19T12:00:00Z',
-        },
-        expectedText: [UI_MESSAGES.CLOUD_SYNC.LAST_REMOTE_PREFIX],
-        absentText: [
-          UI_MESSAGES.CLOUD_SYNC.NEVER_UPLOADED,
-          UI_MESSAGES.CLOUD_SYNC.LAST_UPLOAD_PREFIX,
-        ],
-      },
-      {
-        name: 'lastSuccessfulUploadAt 優先於 lastKnownRemoteUpdatedAt',
-        metadata: {
-          connectionEmail: 'priority@test.dev',
-          lastSuccessfulUploadAt: '2026-04-21T09:00:00Z',
-          lastKnownRemoteUpdatedAt: '2026-04-19T12:00:00Z',
-        },
-        expectedText: [UI_MESSAGES.CLOUD_SYNC.LAST_UPLOAD_PREFIX],
-        absentText: [UI_MESSAGES.CLOUD_SYNC.LAST_REMOTE_PREFIX],
-      },
-    ])('$name', ({ metadata, expectedText, absentText }) => {
-      expect.hasAssertions();
-      renderCloudSyncCard(metadata);
-
-      const text = document.querySelector('#drive-last-upload-text').textContent;
-      for (const expected of expectedText) {
-        expect(text).toContain(expected);
-      }
-      for (const absent of absentText) {
-        expect(text).not.toContain(absent);
-      }
-    });
-
-    it('兩者皆無 → 維持「尚未上載」', () => {
-      renderCloudSyncCard({ connectionEmail: 'nothing@test.dev' });
-      expect(document.querySelector('#drive-last-upload-text').textContent).toBe(
-        UI_MESSAGES.CLOUD_SYNC.NEVER_UPLOADED
-      );
-    });
-
-    it('renders frequency and auto sync status properly', () => {
-      renderCloudSyncCard({
-        connectionEmail: 'test@notion.so',
-        frequency: 'daily',
-        needsManualReview: true,
-      });
-      expect(document.querySelector('#drive-frequency-select').value).toBe('daily');
-      expect(document.querySelector('#drive-auto-sync-status').classList.contains('hidden')).toBe(
-        false
-      );
-      expect(document.querySelector('#drive-auto-sync-status-text').textContent).toBe(
-        UI_MESSAGES.CLOUD_SYNC.AUTO_SYNC_NEEDS_REVIEW
-      );
-
-      // off status hides the container
-      renderCloudSyncCard({
-        connectionEmail: 'test@notion.so',
-        frequency: 'off',
-      });
-      expect(document.querySelector('#drive-auto-sync-status').classList.contains('hidden')).toBe(
-        true
-      );
-    });
-
-    it('hides auto sync status container when frequency is active but no review is needed', () => {
-      // 先觸發 needsManualReview=true 使 container 顯示
-      renderCloudSyncCard({
-        connectionEmail: 'test@notion.so',
-        frequency: 'weekly',
-        needsManualReview: true,
-      });
-      expect(document.querySelector('#drive-auto-sync-status').classList.contains('hidden')).toBe(
-        false
-      );
-
-      // 再切換到 needsManualReview=false：container 必須被隱藏，避免 margin-bottom 佔用空白
-      renderCloudSyncCard({
-        connectionEmail: 'test@notion.so',
-        frequency: 'weekly',
-        needsManualReview: false,
-      });
-      expect(document.querySelector('#drive-auto-sync-status').classList.contains('hidden')).toBe(
-        true
-      );
-      expect(document.querySelector('#drive-auto-sync-status-text').textContent).toBe('');
-    });
   });
 
   describe('initCloudSyncController', () => {
@@ -1350,370 +896,6 @@ describe('DriveCloudSyncController', () => {
         action: RUNTIME_ACTIONS.DRIVE_SYNC_MANUAL_UPLOAD,
         force: true,
       });
-    });
-  });
-
-  describe('refreshCloudSyncCard', () => {
-    it('fetches metadata and re-renders', async () => {
-      driveClient.getDriveSyncMetadata.mockResolvedValue({ connectionEmail: 'mock@a.com' });
-      await refreshCloudSyncCard();
-      expect(document.querySelector('#drive-connected-email').textContent).toBe('mock@a.com');
-    });
-
-    it('can sync remote connection state before rendering', async () => {
-      driveClient.fetchDriveConnectionStatus.mockResolvedValue({
-        connected: true,
-        email: 'fresh@a.com',
-        connectedAt: REMOTE_CONNECTED_AT,
-      });
-      driveClient.getDriveSyncMetadata.mockResolvedValue({
-        connectionEmail: 'fresh@a.com',
-        connectedAt: REMOTE_CONNECTED_AT,
-      });
-
-      await refreshCloudSyncCard({ syncRemote: true });
-
-      expect(driveClient.fetchDriveConnectionStatus).toHaveBeenCalled();
-      expect(driveClient.setDriveConnection).toHaveBeenCalledWith(
-        {
-          email: 'fresh@a.com',
-          connectedAt: REMOTE_CONNECTED_AT,
-        },
-        expect.objectContaining({ resetConflicts: false })
-      );
-      expect(document.querySelector('#drive-connected-email').textContent).toBe('fresh@a.com');
-    });
-
-    it('surfaces identity initialization failure separately during remote connection sync', async () => {
-      const identityError = new Error('IDENTITY_ERROR');
-      driveClient.fetchDriveConnectionStatus.mockResolvedValue({
-        connected: true,
-        email: 'fresh@a.com',
-        connectedAt: REMOTE_CONNECTED_AT,
-      });
-      driveClient.ensureDriveSyncIdentity.mockRejectedValueOnce(identityError);
-      driveClient.getDriveSyncMetadata.mockResolvedValue({ connectionEmail: null });
-
-      await refreshCloudSyncCard({ syncRemote: true });
-
-      expect(loggerErrorSpy).toHaveBeenCalledWith(
-        '[CloudSync] Drive Sync identity initialization failed during connection sync',
-        expect.objectContaining({
-          action: 'syncRemoteDriveConnection',
-          error: sanitizeApiError(identityError, 'drive_connection_identity_init'),
-        })
-      );
-      expect(document.querySelector('#drive-sync-status').textContent).toContain(
-        UI_MESSAGES.CLOUD_SYNC.SYNC_FAILED_PREFIX
-      );
-      expect(document.querySelector('#drive-sync-status').className).toContain('error');
-      expect(driveClient.setDriveConnection).not.toHaveBeenCalled();
-    });
-
-    it('clears local metadata when remote state reports disconnected', async () => {
-      driveClient.fetchDriveConnectionStatus.mockResolvedValue({
-        ...DEFAULT_DISCONNECTED_CONNECTION,
-      });
-      driveClient.getDriveSyncMetadata.mockResolvedValue({ connectionEmail: null });
-
-      await refreshCloudSyncCard({ syncRemote: true });
-
-      expect(driveClient.clearDriveSyncMetadata).toHaveBeenCalled();
-      expect(document.querySelector('#drive-state-disconnected').classList.contains('hidden')).toBe(
-        false
-      );
-    });
-
-    it('syncRemote:true 時同時查詢 snapshot status 並恢復雲端備份顯示', async () => {
-      driveClient.fetchDriveConnectionStatus.mockResolvedValue({
-        connected: true,
-        email: 'refresh@test.dev',
-        connectedAt: '2026-04-21T00:00:00Z',
-      });
-      driveClient.fetchDriveSnapshotStatus.mockResolvedValueOnce({
-        exists: true,
-        updatedAt: REMOTE_SNAPSHOT_UPDATED_AT,
-      });
-      const setSnapshotSpy = jest
-        .spyOn(driveClient, 'setLastKnownRemoteUpdatedAt')
-        .mockResolvedValue();
-      driveClient.getDriveSyncMetadata.mockResolvedValue({
-        connectionEmail: 'refresh@test.dev',
-        lastKnownRemoteUpdatedAt: REMOTE_SNAPSHOT_UPDATED_AT,
-      });
-
-      await refreshCloudSyncCard({ syncRemote: true });
-
-      expect(setSnapshotSpy).toHaveBeenCalledWith(REMOTE_SNAPSHOT_UPDATED_AT);
-      expect(document.querySelector('#drive-last-upload-text').textContent).toContain(
-        UI_MESSAGES.CLOUD_SYNC.LAST_REMOTE_PREFIX
-      );
-    });
-
-    it('clears temporary success status message after timeout', async () => {
-      driveClient.getDriveSyncMetadata
-        .mockResolvedValueOnce({})
-        .mockResolvedValue({ connectionEmail: 'done@test.dev' });
-      driveClient.fetchDriveSnapshotStatus.mockResolvedValue({
-        exists: false,
-        updatedAt: null,
-        size: null,
-        sourceInstallationId: null,
-        sourceProfileId: null,
-      });
-      mockSendMessage.mockResolvedValueOnce({ success: true });
-
-      await initCloudSyncController(true);
-
-      document.querySelector('#drive-upload-button').click();
-      // 等待 preflight + sendMessage 完成
-      await flushAsyncWork();
-
-      expect(document.querySelector('#drive-sync-status').textContent).toContain(
-        UI_MESSAGES.CLOUD_SYNC.UPLOAD_SUCCESS
-      );
-
-      await jest.advanceTimersByTimeAsync(4000);
-
-      expect(document.querySelector('#drive-sync-status').textContent).toBe('');
-      expect(document.querySelector('#drive-sync-status').className).toBe('status-message');
-    });
-
-    it('clears prior status timeout before scheduling a new one', async () => {
-      driveClient.getDriveSyncMetadata
-        .mockResolvedValueOnce({})
-        .mockResolvedValueOnce({ connectionEmail: 'done@test.dev' });
-      mockSendMessage.mockResolvedValueOnce({ success: true });
-
-      await initCloudSyncController(true);
-
-      document.querySelector('#drive-upload-button').click();
-      await flushAsyncWork();
-      expect(document.querySelector('#drive-sync-status').textContent).toContain(
-        UI_MESSAGES.CLOUD_SYNC.UPLOAD_SUCCESS
-      );
-
-      document.querySelector('#drive-disconnect-button').click();
-      await flushAsyncWork();
-      expect(document.querySelector('#drive-sync-status').textContent).toContain(
-        UI_MESSAGES.CLOUD_SYNC.DISCONNECT_SUCCESS
-      );
-
-      await jest.advanceTimersByTimeAsync(3999);
-      expect(document.querySelector('#drive-sync-status').textContent).toContain(
-        UI_MESSAGES.CLOUD_SYNC.DISCONNECT_SUCCESS
-      );
-
-      await jest.advanceTimersByTimeAsync(1);
-      expect(document.querySelector('#drive-sync-status').textContent).toBe('');
-    });
-  });
-  describe('Source Warning (_updateSourceWarning via renderCloudSyncCard)', () => {
-    it('sourceInstallationId 與 installationId 不同時顯示 warning', () => {
-      renderCloudSyncCard(
-        { connectionEmail: 'user@test.dev', installationId: LOCAL_INSTALLATION_ID },
-        { snapshotStatus: { sourceInstallationId: REMOTE_INSTALLATION_ID } }
-      );
-      const warning = document.querySelector('#drive-source-warning');
-      expect(warning.hidden).toBe(false);
-      expect(warning.textContent).toContain('⚠️');
-    });
-
-    it('sourceInstallationId === installationId 時隱藏 warning', () => {
-      renderCloudSyncCard(
-        { connectionEmail: 'user@test.dev', installationId: 'same-id' },
-        { snapshotStatus: { sourceInstallationId: 'same-id' } }
-      );
-      expect(document.querySelector('#drive-source-warning').hidden).toBe(true);
-    });
-
-    it('sourceInstallationId 為 null 時隱藏 warning', () => {
-      renderCloudSyncCard(
-        { connectionEmail: 'user@test.dev', installationId: LOCAL_INSTALLATION_ID },
-        { snapshotStatus: { sourceInstallationId: null } }
-      );
-      expect(document.querySelector('#drive-source-warning').hidden).toBe(true);
-    });
-
-    it('installationId 為 null 時隱藏 warning', () => {
-      renderCloudSyncCard(
-        { connectionEmail: 'user@test.dev', installationId: null },
-        { snapshotStatus: { sourceInstallationId: REMOTE_INSTALLATION_ID } }
-      );
-      expect(document.querySelector('#drive-source-warning').hidden).toBe(true);
-    });
-
-    it('disconnected 狀態時 warning 應隱藏', () => {
-      // 先讓 warning 顯示
-      renderCloudSyncCard(
-        { connectionEmail: 'user@test.dev', installationId: LOCAL_INSTALLATION_ID },
-        { snapshotStatus: { sourceInstallationId: REMOTE_INSTALLATION_ID } }
-      );
-      expect(document.querySelector('#drive-source-warning').hidden).toBe(false);
-
-      // 切換到 disconnected
-      renderCloudSyncCard({ connectionEmail: null });
-      expect(document.querySelector('#drive-source-warning').hidden).toBe(true);
-    });
-  });
-
-  describe('Upload preflight cross-install confirm', () => {
-    beforeEach(async () => {
-      // 基礎連線狀態：已連線
-      driveClient.getDriveSyncMetadata.mockResolvedValue({
-        connectionEmail: 'user@test.dev',
-        installationId: LOCAL_INSTALLATION_ID,
-      });
-      await initCloudSyncController(true);
-    });
-
-    it.each([
-      {
-        name: '偵測跨安裝且使用者取消時，不應送出 DRIVE_SYNC_MANUAL_UPLOAD',
-        confirmResult: false,
-        expectUpload: false,
-      },
-      {
-        name: '偵測跨安裝且使用者確認時，應送出 DRIVE_SYNC_MANUAL_UPLOAD',
-        confirmResult: true,
-        expectUpload: true,
-      },
-    ])('$name', async ({ confirmResult, expectUpload }) => {
-      expect.hasAssertions();
-      driveClient.fetchDriveSnapshotStatus.mockResolvedValueOnce({ ...CROSS_INSTALL_SNAPSHOT });
-      getConfirmDialogMock().mockResolvedValueOnce(confirmResult);
-      mockSendMessage.mockResolvedValueOnce({ success: true });
-
-      document.querySelector('#drive-upload-button').click();
-      await flushAsyncWork();
-
-      const uploadMessage = expect.objectContaining({
-        action: RUNTIME_ACTIONS.DRIVE_SYNC_MANUAL_UPLOAD,
-      });
-      if (expectUpload) {
-        expect(mockSendMessage).toHaveBeenCalledWith(
-          expect.objectContaining({
-            action: RUNTIME_ACTIONS.DRIVE_SYNC_MANUAL_UPLOAD,
-            force: false,
-          })
-        );
-      } else {
-        expect(mockSendMessage).not.toHaveBeenCalledWith(uploadMessage);
-      }
-    });
-
-    it('metadata 缺 installation id 時 preflight 會先建立 identity 再判斷跨安裝', async () => {
-      driveClient.getDriveSyncMetadata.mockResolvedValue({
-        connectionEmail: 'user@test.dev',
-        installationId: null,
-      });
-      driveClient.ensureDriveSyncIdentity.mockResolvedValue(LOCAL_INSTALLATION_ID);
-      driveClient.fetchDriveSnapshotStatus.mockResolvedValueOnce({ ...CROSS_INSTALL_SNAPSHOT });
-      getConfirmDialogMock().mockResolvedValueOnce(false);
-
-      document.querySelector('#drive-upload-button').click();
-      await flushAsyncWork();
-
-      expect(driveClient.ensureDriveSyncIdentity).toHaveBeenCalled();
-      expect(getConfirmDialogMock()).toHaveBeenCalledWith({
-        title: UI_MESSAGES.CLOUD_SYNC.CONFIRM_CROSS_INSTALL_UPLOAD_TITLE,
-        message: UI_MESSAGES.CLOUD_SYNC.CONFIRM_CROSS_INSTALL_UPLOAD,
-        confirmLabel: UI_MESSAGES.CLOUD_SYNC.CONFIRM_CROSS_INSTALL_UPLOAD_OK,
-        cancelLabel: UI_MESSAGES.CLOUD_SYNC.CONFIRM_CROSS_INSTALL_UPLOAD_CANCEL,
-        danger: true,
-      });
-      expect(mockSendMessage).not.toHaveBeenCalledWith(
-        expect.objectContaining({ action: RUNTIME_ACTIONS.DRIVE_SYNC_MANUAL_UPLOAD })
-      );
-    });
-
-    it('preflight 確認期間應先禁用按鈕，取消後恢復互動', async () => {
-      driveClient.fetchDriveSnapshotStatus.mockResolvedValueOnce({ ...CROSS_INSTALL_SNAPSHOT });
-
-      const uploadBtn = document.querySelector('#drive-upload-button');
-      const overlay = document.querySelector('#drive-loading-overlay');
-      let isDisabledDuringConfirm = null;
-      let isOverlayVisibleDuringConfirm = null;
-      getConfirmDialogMock().mockImplementationOnce(() => {
-        isDisabledDuringConfirm = uploadBtn.disabled;
-        isOverlayVisibleDuringConfirm = !overlay.classList.contains('hidden');
-        return Promise.resolve(false);
-      });
-
-      document.querySelector('#drive-upload-button').click();
-      await flushAsyncWork();
-
-      expect(mockSendMessage).not.toHaveBeenCalledWith(
-        expect.objectContaining({ action: RUNTIME_ACTIONS.DRIVE_SYNC_MANUAL_UPLOAD })
-      );
-      expect(isDisabledDuringConfirm).toBe(true);
-      expect(isOverlayVisibleDuringConfirm).toBe(true);
-      expect(uploadBtn.disabled).toBe(false);
-      expect(overlay.classList.contains('hidden')).toBe(true);
-    });
-
-    it('preflight 查詢失敗時，仍應送出 DRIVE_SYNC_MANUAL_UPLOAD（fail-open）', async () => {
-      driveClient.fetchDriveSnapshotStatus.mockRejectedValueOnce(new Error('NETWORK_ERROR'));
-      mockSendMessage.mockResolvedValueOnce({ success: true });
-
-      document.querySelector('#drive-upload-button').click();
-      await flushAsyncWork();
-
-      expect(loggerWarnSpy).toHaveBeenCalledWith(
-        '[CloudSync] Upload preflight check failed, continuing upload',
-        expect.any(Object)
-      );
-      expect(mockSendMessage).toHaveBeenCalledWith(
-        expect.objectContaining({
-          action: RUNTIME_ACTIONS.DRIVE_SYNC_MANUAL_UPLOAD,
-          force: false,
-        })
-      );
-    });
-
-    it('identity 初始化失敗時應阻止 upload 並顯示錯誤', async () => {
-      driveClient.fetchDriveSnapshotStatus.mockResolvedValueOnce({ ...CROSS_INSTALL_SNAPSHOT });
-      driveClient.ensureDriveSyncIdentity.mockRejectedValueOnce(new Error('IDENTITY_ERROR'));
-      mockSendMessage.mockResolvedValueOnce({ success: true });
-
-      document.querySelector('#drive-upload-button').click();
-      await flushAsyncWork();
-
-      expect(loggerErrorSpy).toHaveBeenCalledWith(
-        '[CloudSync] Drive Sync identity initialization failed',
-        expect.objectContaining({
-          error: sanitizeApiError(new Error('IDENTITY_ERROR'), 'drive_sync_identity_init'),
-        })
-      );
-      expect(document.querySelector('#drive-sync-status').className).toContain('error');
-      expect(mockSendMessage).not.toHaveBeenCalledWith(
-        expect.objectContaining({
-          action: RUNTIME_ACTIONS.DRIVE_SYNC_MANUAL_UPLOAD,
-        })
-      );
-    });
-
-    it('同一安裝 ID 時不顯示 confirm dialog，直接上傳', async () => {
-      driveClient.fetchDriveSnapshotStatus.mockResolvedValueOnce({
-        exists: true,
-        updatedAt: 'time',
-        size: null,
-        sourceInstallationId: LOCAL_INSTALLATION_ID,
-        sourceProfileId: null,
-      });
-      driveClient.ensureDriveSyncIdentity.mockResolvedValueOnce(LOCAL_INSTALLATION_ID);
-      mockSendMessage.mockResolvedValueOnce({ success: true });
-
-      document.querySelector('#drive-upload-button').click();
-      await flushAsyncWork();
-
-      // confirm 只會被 download 的初始 setup 呼叫，不應有 CROSS_INSTALL 訊息
-      expect(getConfirmDialogMock()).not.toHaveBeenCalledWith(
-        expect.objectContaining({ message: UI_MESSAGES.CLOUD_SYNC.CONFIRM_CROSS_INSTALL_UPLOAD })
-      );
-      expect(mockSendMessage).toHaveBeenCalledWith(
-        expect.objectContaining({ action: RUNTIME_ACTIONS.DRIVE_SYNC_MANUAL_UPLOAD })
-      );
     });
   });
 });
